dxmesh

+++ src/xdk/d3d9i/d3d9types.h
@@ -943,7 +943,9 @@ typedef struct _D3DVIEWPORT9 { /* Size=0x18 */
     /* 0x0014 */ float MaxZ;
 } D3DVIEWPORT9;
 
-#define D3DDECL_END() { -1, 0, -1, D3DDECLMETHOD_DEFAULT, D3DDECLUSAGE_POSITION, 0 }
+// clang-format off
+#define D3DDECL_END() { (BYTE)-1, 0, -1, D3DDECLMETHOD_DEFAULT, D3DDECLUSAGE_POSITION, 0 }
+// clang-format on
 
 typedef struct _D3DSWAP_STATUS { /* Size=0xc */
     /* 0x0000 */ UINT Flags;
